Enable sanitizers with Noa (#725)

+++ src/runtime/include/sourcemeta/jsonbinpack/runtime_decoder.h
@@ -130,9 +130,10 @@ public:
       // We trust the encoder that the data we are seeing
       // corresponds to a valid 64-bit signed integer.
       return sourcemeta::jsontoolkit::JSON{static_cast<std::int64_t>(
-          -(static_cast<std::int64_t>(this->get_varint() *
-                                      options.multiplier)) +
-          closest_maximum_multiple)};
+          -(static_cast<std::int64_t>(
+              this->get_varint() *
+              static_cast<std::int64_t>(options.multiplier))) +
+          static_cast<std::int64_t>(closest_maximum_multiple))};
     } else {
       const std::uint64_t closest_maximum_multiple{
           static_cast<std::uint32_t>(std::abs(closest_maximum)) *
@@ -140,9 +141,10 @@ public:
       // We trust the encoder that the data we are seeing
       // corresponds to a valid 64-bit signed integer.
       return sourcemeta::jsontoolkit::JSON{static_cast<std::int64_t>(
-          -(static_cast<std::int64_t>(this->get_varint() *
-                                      options.multiplier)) -
-          closest_maximum_multiple)};
+          -(static_cast<std::int64_t>(
+              this->get_varint() *
+              static_cast<std::int64_t>(options.multiplier))) -
+          static_cast<std::int64_t>(closest_maximum_multiple))};
     }
   }
 
@@ -153,7 +155,8 @@ public:
     // We trust the encoder that the data we are seeing
     // corresponds to a valid 64-bit signed integer.
     return sourcemeta::jsontoolkit::JSON{static_cast<std::int64_t>(
-        this->get_varint_zigzag() * options.multiplier)};
+        this->get_varint_zigzag() *
+        static_cast<std::int64_t>(options.multiplier))};
   }
 
   /// @}
